Extraction of translatable strings and binaries.
This commit adds string/binary extraction with the help of a
parse_transform and the ability to output simple PO-files.
On top of this extracting parse_transform I've built some helper
functions to make it easier to work with and call from the command-line.

There is the ability to extract references and comments but lots of
other options that is provided by xgettext are so far not supported
but I hope that I've made it easy to extend in the future.